// project/src/models/eft/common/tables/IRepeatableQuests.ts
export interface IReward 
{
    index: number
    type: string
    value: number
    target?: string
    items?: Item[]
}

export interface IRepeatableQuestDatabase 
{
    templates: ITemplates
    rewards: IRewardOptions
    data: IOptions
    samples: ISampleQuests[]
}
  
export interface ITemplates 
{
    Elimination: IRepeatableQuest
    Completion: IRepeatableQuest
    Exploration: IRepeatableQuest
}

export interface IPmcDataRepeatableQuest 
{
    id?: string
    name: string
    activeQuests: IRepeatableQuest[]
    inactiveQuests: IRepeatableQuest[]
    endTime: number
    changeRequirement: TChangeRequirementRecord // what it costs to reset <QuestId, ChangeRequirement> redundant to change requirements within the IRepeatableQuest
}

export type TChangeRequirementRecord = Record<string, IChangeRequirement>

export interface IChangeRequirement
{
    changeCost: IChangeCost[]
    changeStandingCost: number 
}
  
export interface IChangeCost
{
    templateId: string // what item it will take to reset daily
    count: number // amount of item needed to reset
}

export interface IRepeatableQuest 
{
    _id: any
    traderId: string
    location: any
    image: string
    type: string
    isKey: boolean
    restartable: boolean
    instantComplete: boolean
    secretQuest: boolean
    canShowNotificationsInGame: boolean
    rewards: IRewards
    conditions: IConditions
    side: string
    name: string
    note: string
    description: string
    successMessageText: string
    failMessageText: string
    startedMessageText: string
    changeQuestMessageText: string
    templateId: string
    changeCost: IChangeCost[]
    changeStandingCost: number
}

export interface IRewards 
{
    Started: IReward[]
    Success: IReward[]
    Fail: IReward[]
}

export interface IConditions 
{
    AvailableForStart: any[]
    AvailableForFinish: IAvailableFor[]
    Fail: any[]
}

export interface IAvailableFor 
{
    _props: IAvailableForProps
    _parent: string
    dynamicLocale: boolean
}