// project/src/models/enums/HideoutAreas.ts
export enum HideoutAreas 
    {
    NOTSET = -1,
    VENTS = 0,
    SECURITY = 1,
    LAVATORY = 2,
    STASH = 3,
    GENERATOR = 4,
    HEATING = 5,
    WATER_COLLECTOR = 6,
    MEDSTATION = 7,
    NUTRITION_UNIT = 8,
    REST_SPACE = 9,
    WORKBENCH = 10,
    INTEL_CENTER = 11,
    SHOOTING_RANGE = 12,
    LIBRARY = 13,
    SCAV_CASE = 14,
    ILLUMINATION = 15,
    PLACE_OF_FAME = 16,
    AIR_FILTERING = 17,
    SOLAR_POWER = 18,
    BOOZE_GENERATOR = 19,
    BITCOIN_FARM = 20,
    CHRISTMAS_TREE = 21,
    EMERGENCY_WALL = 22,
    GYM = 23
}